add doc_comments
