+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 
 members = [
-  # "clearned_example",
+  "clearned_example",
   "limousine_core",
   # "clearned_bench",
   "limousine_derive",
