+++ limousine_core/src/lib.rs
@@ -1,23 +1,37 @@
-use bytemuck::{Pod, Zeroable};
-use mmap_buffer::Buffer;
+//! This crate contains building blocks for building hybrid index data structures.
+//!
+//! This crate should not be imported directly, unless one needs to build a very custom structure
+//! for benchmarking or debugging purposes. The engine should be accessed through the
+//! [`limousine_engine`](https://crates.io/crates/limousine_engine) crate.
+
+#![deny(missing_docs)]
+
+use bytemuck::Pod;
 use num::PrimInt;
 use std::{
     borrow::Borrow,
     fmt::Debug,
-    ops::Deref,
     path::{Path, PathBuf},
 };
 use trait_set::trait_set;
 
+mod base;
 mod classical;
+mod entry;
 mod hybrid;
 mod learned;
-pub mod search;
+mod search;
 
-pub use classical::BTreeLayer;
+#[cfg(feature = "standard")]
+/// Common interface implementations of standard index structures
+/// such as BTreeMap, which can be useful for testing/benchmarking.
+pub mod standard;
+
 pub use hybrid::HybridIndex;
 pub use hybrid::HybridIndexRangeIterator;
 pub use hybrid::HybridLayer;
+
+pub use classical::BTreeLayer;
 pub use learned::pgm_node::PGMLayer;
 
 /// Generic error type (to avoid a dependency on anyhow)
@@ -32,7 +46,7 @@ trait_set! {
     pub trait Key = Value + PrimInt;
 }
 
-/// Represents an immutable tree index over a sorted slice of entries
+/// An immutable tree index which stores a sorted slice of entries.
 pub trait ImmutableIndex<K: Key, V: Value>: Sized {
     /// Build an index over the given data in memory
     /// This method assumes the data is sorted, and without key repetitions
@@ -55,6 +69,7 @@ pub trait ImmutableIndex<K: Key, V: Value>: Sized {
     /// Returns an iterator which iterates the entries between `low` and `high`, inclusive
     fn range(&self, low: &K, high: &K) -> Self::RangeIterator<'_>;
 
+    /// Iterator type returned by range operations
     type RangeIterator<'e>: Iterator<Item = (&'e K, &'e V)>
     where
         Self: 'e;
@@ -62,12 +77,14 @@ pub trait ImmutableIndex<K: Key, V: Value>: Sized {
 
 /// Some layer of nodes on top of which an index layer can be built.
 pub trait NodeLayer<K: Key> {
+    /// Node type of the layer
     type Node: Borrow<K>;
 
     /// Returns a slice of keys, each of which represent the
     /// minimum keys of the underlying node.
     fn nodes(&self) -> &[Self::Node];
 
+    /// Number of nodes in the layer
     fn len(&self) -> usize {
         self.nodes().len()
     }
@@ -85,6 +102,9 @@ pub struct ApproxPos {
 /// the full flexibility of the `ApproxPos` generality,
 /// but since we must support learned index layers, we need this generality.
 pub trait InternalLayer<K: Key>: NodeLayer<K> {
+    /// Search method for a layer, takes a key and a range of nodes where the key
+    /// might be found, and returns a range of nodes where the key might be found in
+    /// the next layer.
     fn search(&self, key: &K, range: ApproxPos) -> ApproxPos;
 }
 
@@ -109,81 +129,6 @@ pub trait InternalLayerBuild<K: Key>: NodeLayer<K> {
     fn load(path: impl AsRef<Path>) -> Result<Self>
     where
         Self: Sized;
-}
-
-// ------------------------------------------------------
-// Entry
-// ------------------------------------------------------
-
-#[derive(Clone, Copy, Debug)]
-#[repr(C)]
-pub struct Entry<K, V> {
-    pub key: K,
-    pub value: V,
-}
-
-impl<K, V> Entry<K, V> {
-    pub fn new(key: K, value: V) -> Self {
-        Self { key, value }
-    }
-}
-
-impl<K, V> Borrow<K> for Entry<K, V> {
-    fn borrow(&self) -> &K {
-        &self.key
-    }
-}
-
-unsafe impl<K: Zeroable, V: Zeroable> Zeroable for Entry<K, V> {}
-unsafe impl<K: Pod, V: Pod> Pod for Entry<K, V> {}
-
-// ------------------------------------------------------
-// Base Layer
-// ------------------------------------------------------
-
-pub struct BaseLayer<K: Key, V: Value> {
-    data: Buffer<Entry<K, V>>,
-}
-
-impl<K: Key, V: Value> BaseLayer<K, V> {
-    pub fn build(base: impl ExactSizeIterator<Item = (K, V)>) -> Self {
-        Self {
-            data: Buffer::from_vec_in_memory(base.map(|(k, v)| Entry::new(k, v)).collect()),
-        }
-    }
-
-    pub fn build_disk(
-        base: impl ExactSizeIterator<Item = (K, V)>,
-        path: impl AsRef<Path>,
-    ) -> Result<Self> {
-        let data: Vec<Entry<K, V>> = base.map(|(k, v)| Entry::new(k, v)).collect();
-
-        Ok(Self {
-            data: Buffer::from_slice_on_disk(&data[..], path)?,
-        })
-    }
-
-    pub fn load(path: impl AsRef<Path>) -> Result<Self> {
-        Ok(Self {
-            data: Buffer::load_from_disk(path)?,
-        })
-    }
-}
-
-impl<K: Key, V: Value> NodeLayer<K> for BaseLayer<K, V> {
-    type Node = Entry<K, V>;
-
-    fn nodes(&self) -> &[Self::Node] {
-        self.data.deref()
-    }
-}
-
-impl<K: Key, V: Value> Deref for BaseLayer<K, V> {
-    type Target = [Entry<K, V>];
-
-    fn deref(&self) -> &Self::Target {
-        self.data.deref()
-    }
 }
 
 // ---------------------------------------------------------------------------
